Refactor: Error Name Exception Change
