+++ src/main/kotlin/com/example/ggyunispring/config/security/SecurityConfig.kt
@@ -2,9 +2,8 @@ package com.example.ggyunispring.config.security
 
 import com.example.ggyunispring.common.jwt.JwtAuthenticationFilter
 import com.example.ggyunispring.common.jwt.JwtProvider
-import com.example.ggyunispring.dto.response.ResponseDTO
-import com.example.ggyunispring.error.ExceptionResponse
-import com.example.ggyunispring.error.ExceptionType
+import com.example.ggyunispring.exception.ExceptionResponse
+import com.example.ggyunispring.exception.ExceptionType
 import com.fasterxml.jackson.databind.ObjectMapper
 import org.springframework.http.HttpMethod
 import org.springframework.http.HttpStatus
