+++ src/main/kotlin/com/example/ggyunispring/exception/ExceptionResponse.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.error
+package com.example.ggyunispring.exception
 
 data class ExceptionResponse(
     val status: Int,
