+++ src/main/kotlin/com/example/ggyunispring/exception/member/GoogleIdTokenException.kt
@@ -1,4 +1,4 @@
-package com.example.ggyunispring.error.member
+package com.example.ggyunispring.exception.member
 
 class GoogleIdTokenException : RuntimeException() {
 }
